requirements.txt
streamlit==1.33.0
langchain==0.2.14
langchain-core==0.2.32
langchain-community==0.2.12
langchain-groq==0.1.5
chromadb==0.4.24
sentence-transformers==2.6.1
pymupdf==1.24.1
python-dotenv==1.0.1